// src/api/Dictionary.ts
export default interface Dictionary {
	BUTTON_NEXT: string;
	BUTTON_SUBMIT:  string;
	BUTTON_EDIT:  string;
	BUTTON_CANCEL:  string;
	BUTTON_SAVE:  string;
	BUTTON_ADD_NOTE:  string;
	BUTTON_SWITCH_LANGUAGES:  string;
	CONFIRMATION_PHRASE_TYPE:  string;
	CONFIRMATION_CZ_VERB_ASPECT:  string;
	CONFIRMATION_CZ_GENDER:  string;
	CONFIRMATION_WORD_TYPE:  string;
	CZ_GENDER_SELECT_LABEL:  string;
	CZ_GENDER_OPTION_NEUTER:  string;
	CZ_GENDER_OPTION_FEMININE:  string;
	CZ_GENDER_OPTION_MASCULINE:  string;
	CZ_GENDER_OPTION_MASCULINE_ANIMATUM:  string;
	CZ_VERB_ASPECT_SELECT_LABEL:  string;
	CZ_VERB_ASPECT_OPTION_PERFECTIVE:  string;
	CZ_VERB_ASPECT_OPTION_IMPERFECTIVE:  string;
	MENULABEL_ADD:  string;
	MENULABEL_HOME:  string;
	MENULABEL_SETTINGS:  string;
	PHRASE_TYPE_OPTION_COLLOQUIALISM:  string;
	PHRASE_TYPE_OPTION_IDIOM:  string;
	PHRASE_TYPE_OPTION_OTHER:  string;
	PHRASE_TYPE_OPTION_PROVERB:  string;
	PHRASE_TYPE_SELECT_LABEL:  string;
	PLACEHOLDER_INPUT_IN_CZECH:  string;
	PLACEHOLDER_INPUT_IN_ENGLISH:  string;
	PLACEHOLDER_INPUT_NO_LANGUAGE:  string;
	PLACEHOLDER_TRANSLATION_IN_CZECH:  string;
	PLACEHOLDER_TRANSLATION_IN_ENGLISH:  string;
	SETTINGS_LANGUAGE_OPTION_CZ:  string;
	SETTINGS_LANGUAGE_OPTION_EN:  string;
	SETTINGS_LANGUAGE_OPTION_NONE:  string;
	SETTINGS_SECTION_DEFAULT_INPUT_LANG:  string;
	SETTINGS_SECTION_UI_LANGUAGE:  string;
	SELECT_LABEL_PAIRING_NOTES: string;
	SELECT_LABEL_LEXEME_NOTES: string;
	TAB_ENGLISH:  string;
	TAB_CZECH:  string;
	TAB_CONFIRM:  string;
	WHICH_LANGUAGE:  string;
	WORD_TYPE_SELECT_LABEL:  string;
	WORD_TYPE_OPTION_VERB:  string;
	WORD_TYPE_OPTION_NOUN:  string;
	WORD_TYPE_OPTION_ADJECTIVE:  string;
	WORD_TYPE_OPTION_ADVERB:  string;
	WORD_TYPE_OPTION_PRONOUN:  string;
	WORD_TYPE_OPTION_PREPOSITION:  string;
	WORD_TYPE_OPTION_CONJUNCTION:  string;
	WORD_TYPE_OPTION_GERUND:  string;
}